# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dtorkon API",
    "description": "FastAPI backend for the dtorkon mini-blog. The service handles admin-only auth, public/admin posts, site profile editing, analytics, SQLite persistence, searchable publishing, and backend-mediated uploads to Yandex Object Storage over the S3 API.",
    "version": "0.1.0-fastapi"
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "example.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/admin/settings/transcription": {
      "get": {
        "tags": [
          "Admin Settings"
        ],
        "summary": "Read transcription configuration status",
        "operationId": "get_transcription_settings_api_admin_settings_transcription_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionSettingsRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/admin/settings/transcription/groq-api-key": {
      "put": {
        "tags": [
          "Admin Settings"
        ],
        "summary": "Set Groq API key for transcription (stored in SQLite)",
        "operationId": "set_groq_api_key_api_admin_settings_transcription_groq_api_key_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroqApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionSettingsRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Admin Settings"
        ],
        "summary": "Remove stored Groq API key",
        "operationId": "delete_groq_api_key_api_admin_settings_transcription_groq_api_key_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranscriptionSettingsRead": {
        "properties": {
          "groqConfigured": {
            "type": "boolean",
            "title": "Groqconfigured"
          },
          "groqApiBase": {
            "type": "string",
            "title": "Groqapibase"
          },
          "groqSpeechModel": {
            "type": "string",
            "title": "Groqspeechmodel"
          }
        },
        "type": "object",
        "required": [
          "groqConfigured",
          "groqApiBase",
          "groqSpeechModel"
        ],
        "title": "TranscriptionSettingsRead"
      },
      "UpdateGroqApiKeyRequest": {
        "properties": {
          "apiKey": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "title": "UpdateGroqApiKeyRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dtorkon_session"
      }
    }
  },
...
}